change column type of amount

+++ database/migrations/20230214051017_create_init_tables.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE "orders" (
      "order_name" varchar unique not null,
      "customer_name" varchar not null,
      "customer_company" varchar not null,
-     "delivered_amount" integer,
-     "total_amount" integer not null,
+     "delivered_amount" float,
+     "total_amount" float not null,
      "order_date" timestamptz not null default current_timestamp
 );
